+++ server/src/main/java/com/linbit/linstor/core/identifier/SnapshotName.java
@@ -13,11 +13,26 @@ public class SnapshotName extends GenericName
     private static final int MAX_LENGTH = 48;
 
     public static final byte[] VALID_CHARS = {'_'};
+    public static final byte[] VALID_CHARS_INTERNAL =
+    {
+        '_', '.'
+    };
     public static final byte[] VALID_INNER_CHARS = {'_', '-'};
 
     public SnapshotName(String snapshotName) throws InvalidNameException
     {
+        this(snapshotName, false);
+    }
+
+    public SnapshotName(String snapshotName, boolean internal) throws InvalidNameException
+    {
         super(snapshotName);
-        Checks.nameCheck(snapshotName, MIN_LENGTH, MAX_LENGTH, VALID_CHARS, VALID_INNER_CHARS);
+        Checks.nameCheck(
+            snapshotName,
+            MIN_LENGTH,
+            MAX_LENGTH,
+            internal ? VALID_CHARS_INTERNAL : VALID_CHARS,
+            VALID_INNER_CHARS
+        );
     }
 }